+++ packages/api/tests/lib/writeRecords/test-write-granules.js
@@ -742,7 +742,6 @@ test.serial('writeGranulesFromMessage() on re-write saves granule records to Pos
 
   const validNullableGranuleKeys = [
     'beginningDateTime',
-    'active',
     'cmrLink',
     'createdAt',
     'duration',
@@ -1075,7 +1074,6 @@ test.serial('writeGranulesFromMessage() on re-write saves granule records to Pos
 
   const validNullableGranuleKeys = [
     'beginningDateTime',
-    'active',
     'cmrLink',
     'createdAt',
     'duration',
@@ -3693,7 +3691,6 @@ test.serial('writeGranuleFromApi() given a granule with all fields populated is 
 
   const validNullableGranuleKeys = [
     'beginningDateTime',
-    'active',
     'cmrLink',
     'duration',
     'endingDateTime',