+++ packages/db/src/migrations/20250902134823_granules_add_active.ts
@@ -4,11 +4,15 @@ export const up = async (knex: Knex): Promise<void> => {
   if (!await knex.schema.hasColumn('granules', 'active')) {
     await knex.schema.table('granules', (table) => {
       table
-        .boolean('active')
-        .comment('Active Status');
+        .string('active', 1)
+        .comment('Flag for active Status');
     });
 
-    await knex('granules').update('active', true);
+    await knex('granules').update('active', 'A');
+
+    await knex.schema.table('granules', (table) => {
+      table.text('active').notNullable().alter();
+    });
   }
 
   await knex.raw('CREATE INDEX CONCURRENTLY IF NOT EXISTS granules_active_index ON granules(active)');