+++ packages/db/src/types/granule.ts
@@ -6,6 +6,7 @@ export interface PostgresGranuleUniqueColumns {
 export interface PostgresGranule extends PostgresGranuleUniqueColumns {
   archived: boolean,
   producer_granule_id: string,
+  active: string,
   status?: GranuleStatus,
   cmr_link?: string | null,
   error?: object | null,
@@ -28,7 +29,6 @@ export interface PostgresGranule extends PostgresGranuleUniqueColumns {
   processing_start_date_time?: Date | null,
   processing_end_date_time?: Date | null,
   query_fields?: unknown | null,
-  active?: boolean | null,
 }
 
 // product_volume is stored as a BigInt in Postgres. It returns from PG to Node