+++ CHANGELOG.md
@@ -18,7 +18,7 @@ Please follow the instructions before upgrading Cumulus
 ### Added
 
 - **CUMULUS-4146**
-  - Added new  column `active` to Postgres `granules` table.
+  - Added new column `active` to Postgres `granules` table to discern whether a duplicate granule is active or hidden
   - Added `active` property to `granule` record schema.
   - Created migration for `active` column and index for granules table